requirements.txt
confluent_kafka==2.3.0
requests==2.31.0
kafka-python-ng==2.2.2